feat(lib-common): promote hat shapeAdjustments from app-vscode
Moves the shape-adjustment constants (defaultShapeAdjustments,
DEFAULT_HAT_HEIGHT_EM, DEFAULT_VERTICAL_OFFSET_EM, HatAdjustments,
IndividualHatAdjustmentMap) out of app-vscode into lib-common, alongside
the already-shared hatStyles.types, so non-VS-Code consumers (e.g.
@cursorless/command-visualizer) can import them instead of vendoring a copy.

app-vscode's original shapeAdjustments.ts becomes a backward-compatible
re-export shim, so its consumers (VscodeHatRenderer,
performPr1868ShapeUpdateInit, the hatAdjustments scripts) are unchanged.
No behavior change — values are byte-identical to the previous location.
